+++ packages/badge/src/Badge.js
@@ -15,18 +15,19 @@ type Props = {
   type?: BadgeTypes,
   className?: string,
   style?: object,
-  children?: any
+  children?: any,
+  ariaLabel?: string
 };
 
 const Badge = (props: Props) => {
-  const { type, className, style = {}, children } = props
+  const { type, className, style = {}, children,  ariaLabel } = props
   const badgeClass = classNames(
     'a-badge',
     className,
     { [`${typeClasses[type]}`]: !!type },
   );
 
-  return <span className={badgeClass} style={style}>{children}</span>
+  return <span className={badgeClass} title={ariaLabel}  aria-label={ariaLabel} style={style}>{children}</span>
 }
 
 export default Badge;